+++ smasher/__init__.py
@@ -1,0 +1,7 @@
+__title__      = "smasher"
+__version__    = "0.1.0"
+__author__     = "Fox Peterson"
+__maintainer__ = "Fox Peterson"
+__email__      = "fox@tinybike.net"
+
+from .smasher import smashWorkers, form_connection
