+++ smasher/data/daily_precip_shifted.csv
