+++ smasher/smashCreateTable.py
