+++ smasher/smasher.py
