--- test/.travis.yml
@@ -1,16 +1,0 @@
-language: python
-notifications:
-  email: true
-python:
-  - "2.6"
-  - "2.7"
-install:
-  - "sudo apt-get install freetds-dev python-numpy python-matplotlib python-scipy python-pymssql -qq -y"
-  - "easy_install -U setuptools"
-  - "pip install --upgrade pip"
-  - "pip install -r requirements.txt"
-  - "pip install -r test_requirements.txt"
-script:
-  - py.test test/test_smash.py --doctest-modules -v --cov glitch --cov-report term-missing
-after_success:
-  - coveralls
